backend/apps/core/factories.py
import factory

from django.contrib.auth import get_user_model

from faker import Faker


User = get_user_model()

fake = Faker()


class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User

    username = factory.LazyAttribute(lambda x: fake.user_name())
    email = factory.LazyAttribute(lambda x: fake.email())
    password = factory.PostGenerationMethodCall('set_password', 'password')
