requirements.txt
adrf==0.1.8
asgiref==3.8.1
asttokens==3.0.0
async-property==0.2.2
async-timeout==5.0.1
black==24.10.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
coverage==7.6.8
decorator==5.1.1
Django==5.1.3
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
factory_boy==3.3.1
Faker==33.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
locust==2.32.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
